Theo

+++ src/main/java/com/Aster/Service/CustomerService.java
@@ -4,6 +4,7 @@ import com.Aster.Database.CustomerDB;
 import com.Aster.Model.Customer;
 import com.Aster.Model.Order;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.expression.spel.ast.NullLiteral;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -15,7 +16,7 @@ public class CustomerService {
     @Autowired
     public CustomerService(CustomerDB customerDB){this.customerDB = customerDB;}
 
-    public int addCustomer(Customer customer){
+    public int addCustomer(Customer customer) throws Exception{
         return customerDB.addCustomer(customer);
     }
     public String getCustomername(String email) throws Exception{
@@ -27,30 +28,45 @@ public class CustomerService {
 
 
     public int addCart(String email, Order order) throws Exception{
-        return customerDB.addCart(email, order);
+        if(customerDB.isvalid(email)) {
+            return customerDB.addCart(email, order);
+        }
+        else throw new Exception("Email does not exist");
     }
     public List<Order> viewCart(String email) throws Exception{
-        return customerDB.viewCart(email);
+        if(customerDB.isvalid(email)) {
+            return customerDB.viewCart(email);
+        }
+        else throw new Exception("Email does not exist");
     }
     public int removeCart(String email, Order order) throws Exception{
-        return customerDB.removeCart(email, order);
+        if(customerDB.isvalid(email)) {
+            return customerDB.removeCart(email, order);
+        }
+        else throw new Exception("Email does not exist");
     }
     public int emptyCart(String email) throws Exception{
-        return customerDB.emptyCart(email);
+        if(customerDB.isvalid(email)) {
+            return customerDB.emptyCart(email);
+        }
+        else throw new Exception("Email does not exist");
     }
 
 
     public int checkout(String email) throws Exception{
-        //last check of cart
-        customerDB.viewCart(email);
-        //make payment
+        if(customerDB.isvalid(email)) {
+            //last check of cart
+            customerDB.viewCart(email);
+            //make payment
 
-        //send order to florist
+            //send order to florist
 
-        //update history
+            //update history both
 
-        //empty cart
-        customerDB.emptyCart(email);
-        return 0;
+            //empty cart
+            customerDB.emptyCart(email);
+            return 0;
+        }
+        else throw new Exception("Email does not exist");
     }
 }
